refactor: consolidate eigensolve sidecar drivers into backend-agnostic script
Closes #127
