+++ reference/driver/eigensolve_from_sidecar.py
@@ -1,0 +1,233 @@
+"""Backend-agnostic eigensolve driver for cube-cavity sidecars (Epic #88 / issue #127).
+
+Consolidates ``eigensolve_from_tfjava.py`` and ``eigensolve_from_onnx.py`` into a
+single script parameterised by ``--backend``. The two existing scripts are thin
+wrappers that inject ``--backend tfjava`` / ``--backend onnx`` and delegate here,
+so all CI invocations and external callers continue to work without change.
+
+Usage
+=====
+    python3 reference/driver/eigensolve_from_sidecar.py \\
+        path/to/reduced_kM.json \\
+        --backend tfjava|onnx \\
+        [--k 5] [--dense] [--out path/to/eigenresult.json]
+
+The output JSON is in fixture-schema v1 so the harness can compare it to the
+JAX / NumPy baselines without language-specific code paths (see
+``compare_eigenvalues.py``).
+"""
+
+from __future__ import annotations
+
+import argparse
+import json
+import sys
+from pathlib import Path
+
+import numpy as np
+
+# ---------------------------------------------------------------------------
+# Per-backend metadata
+# ---------------------------------------------------------------------------
+
+BACKENDS: dict[str, dict] = {
+    "tfjava": {
+        "fixture_id_suffix": "tfjava_eigensolve",
+        "print_prefix": "Loaded TF-Java sidecar",
+        "description": (
+            "Eigenvalues from the TF-Java assembly + SciPy eigensolve seam. "
+            "Cross-checked against the JAX baseline; see #93."
+        ),
+        "eigenvalue_description": (
+            "Lowest 5 scalar Helmholtz eigenvalues from TF-Java assembly "
+            "+ SciPy eigensolve. Cross-language drift tolerance is 1e-8 "
+            "absolute (consistent with the JAX baseline tolerance)."
+        ),
+        "provenance_source": (
+            "reference/tf_java/cube_cavity (assembly) → "
+            "reference/driver/eigensolve_from_sidecar.py (eigensolve seam)"
+        ),
+        "provenance_issue": "#93",
+    },
+    "onnx": {
+        "fixture_id_suffix": "onnx_eigensolve",
+        "print_prefix": "Loaded ONNX sidecar",
+        "description": (
+            "Eigenvalues from the ONNX assembly + SciPy eigensolve seam. "
+            "Cross-checked against the JAX and NumPy baselines; see Epic "
+            "#88 Phase F.2 (issue #123)."
+        ),
+        "eigenvalue_description": (
+            "Lowest 5 scalar Helmholtz eigenvalues from ONNX assembly "
+            "+ SciPy eigensolve. Cross-language drift tolerance is "
+            "1e-8 absolute (consistent with the JAX baseline tolerance)."
+        ),
+        "provenance_source": (
+            "reference/onnx/cube_cavity (assembly) -> "
+            "reference/driver/eigensolve_from_sidecar.py (eigensolve seam)"
+        ),
+        "provenance_issue": "#123",
+    },
+}
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _flatten_to_array(field, dtype=np.float64):
+    """Per the fixture schema, fields may be nested or flat; this
+    flattens to a 1-D ndarray of the declared dtype."""
+    if isinstance(field, dict):
+        data = field["data"]
+        shape = field["shape"]
+    else:
+        raise ValueError(f"unexpected field shape: {type(field)}")
+    arr = np.asarray(data, dtype=dtype).ravel()
+    expected = int(np.prod(shape))
+    if arr.size != expected:
+        raise ValueError(
+            f"data length {arr.size} does not match shape {shape} (= {expected})"
+        )
+    return arr.reshape(shape)
+
+
+# ---------------------------------------------------------------------------
+# Main
+# ---------------------------------------------------------------------------
+
+
+def main():
+    parser = argparse.ArgumentParser(
+        description="Backend-agnostic SciPy eigensolve over a reduced_kM.json sidecar."
+    )
+    parser.add_argument(
+        "sidecar",
+        help="Path to the reduced_kM.json sidecar produced by the assembly step.",
+    )
+    parser.add_argument(
+        "--backend",
+        required=True,
+        choices=list(BACKENDS),
+        metavar="BACKEND",
+        help="Assembly backend that produced the sidecar. One of: "
+        + ", ".join(BACKENDS),
+    )
+    parser.add_argument(
+        "--k",
+        type=int,
+        default=5,
+        help="Number of lowest eigenmodes to extract.",
+    )
+    parser.add_argument(
+        "--dense",
+        action="store_true",
+        help="Force dense eigh (else auto-select by problem size).",
+    )
+    parser.add_argument("--out", default="eigenresult.json")
+    args = parser.parse_args()
+
+    if args.backend not in BACKENDS:
+        # argparse `choices` already catches this; guard for programmatic use.
+        print(
+            f"Unknown backend '{args.backend}'. Valid options: "
+            + ", ".join(BACKENDS),
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    meta = BACKENDS[args.backend]
+
+    sidecar_path = Path(args.sidecar)
+    if not sidecar_path.exists():
+        print(f"Sidecar not found: {sidecar_path}", file=sys.stderr)
+        sys.exit(2)
+    with open(sidecar_path) as f:
+        sidecar = json.load(f)
+
+    k_int = _flatten_to_array(sidecar["outputs"]["k_int"], dtype=np.float64)
+    m_int = _flatten_to_array(sidecar["outputs"]["m_int"], dtype=np.float64)
+    n_int = k_int.shape[0]
+    if k_int.shape != (n_int, n_int) or m_int.shape != (n_int, n_int):
+        print(
+            f"Expected square matrices, got K {k_int.shape}, M {m_int.shape}",
+            file=sys.stderr,
+        )
+        sys.exit(3)
+
+    n = int(sidecar["inputs"]["n"]["data"][0])
+    side = float(sidecar["inputs"]["side"]["data"][0])
+    print(f"{meta['print_prefix']}: n={n}, side={side}, n_int={n_int}")
+    print(f"  trace(K_int) = {np.trace(k_int):.12e}")
+    print(f"  trace(M_int) = {np.trace(m_int):.12e}")
+
+    if args.dense or n_int < 30:
+        from scipy.linalg import eigh
+        eigvals, eigvecs = eigh(k_int, m_int)
+        eigvals = eigvals[:args.k]
+        eigvecs = eigvecs[:, :args.k]
+        solver = "scipy.linalg.eigh (dense)"
+    else:
+        import scipy.sparse as sp
+        import scipy.sparse.linalg as spla
+        k_sp = sp.csr_matrix(k_int)
+        m_sp = sp.csr_matrix(m_int)
+        eigvals, eigvecs = spla.eigsh(k_sp, k=args.k, M=m_sp, sigma=0.0, which="LM")
+        order = np.argsort(eigvals)
+        eigvals = eigvals[order]
+        eigvecs = eigvecs[:, order]
+        solver = "scipy.sparse.linalg.eigsh (ARPACK, shift-invert sigma=0)"
+
+    print(f"Solver: {solver}")
+    print("Lowest eigenvalues:")
+    for i, lam in enumerate(eigvals):
+        print(f"  λ[{i}] = {lam:.6e}")
+
+    # Build fixture-schema-shaped output for harness comparison.
+    result_fixture = {
+        "schema_version": "1",
+        "fixture_id": f"cube_cavity/n{n}_{meta['fixture_id_suffix']}",
+        "description": meta["description"],
+        "units": "dimensionless",
+        "inputs": {
+            "n": {
+                "shape": [1],
+                "dtype": "i64",
+                "description": "Cells per side.",
+                "data": [n],
+            },
+            "side": {
+                "shape": [1],
+                "dtype": "f64",
+                "description": "Cube side.",
+                "data": [side],
+            },
+        },
+        "outputs": {
+            "eigenvalues": {
+                "shape": [args.k],
+                "dtype": "f64",
+                "description": meta["eigenvalue_description"],
+                "tolerance_abs": 1.0e-8,
+                "data": eigvals.tolist(),
+            },
+        },
+        "provenance": {
+            "source": meta["provenance_source"],
+            "verified_against": (
+                "reference/jax/cube_cavity.py and "
+                "reference/numpy/cube_cavity_minimal.py"
+            ),
+            "issue": meta["provenance_issue"],
+        },
+    }
+    out_path = Path(args.out)
+    out_path.parent.mkdir(parents=True, exist_ok=True)
+    with open(out_path, "w") as f:
+        json.dump(result_fixture, f, indent=2)
+        f.write("\n")
+    print(f"Wrote {out_path}")
+
+
+if __name__ == "__main__":
+    main()
